Bootcamp Ownership

+++ controllers/bootcamps.js
@@ -49,14 +49,22 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route    PUT /api/v1/bootcamps
 // @access   Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true
-    });
+    let bootcamp = await Bootcamp.findById(req.params.id);
 
     if(!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
     }
+
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp`, 401)
+        );
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true
+    });
 
     res.status(200).json({ success: true, data: bootcamp });
 });
@@ -69,6 +77,12 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
 
     if(!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
+    }
+
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp`, 401)
+        );
     }
 
     bootcamp.deleteOne();
@@ -114,6 +128,12 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 
     if(!req.files) {
         return next(new ErrorResponse(`Please upload a file`, 400));
+    }
+
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp`, 401)
+        );
     }
 
     const file = req.files.file;
